New feature: decoupling frame rate of slave processes from the master process.
This allows to run processes for displays with different refresh rates and still
have each process render at its preferred frame rate.

The master process should be the one reaching the highest frame rate. To
guarantee this, it is a good idea to disable synchronization to the display
refresh rate for the master process, or to use a master process without windows
(i.e. doing no rendering at all).

Synchronization to the display refresh rate is now a per-process configuration
option.